+++ src/lib/semantics/analyze.ts
@@ -1,3 +1,4 @@
+import { UguisuError } from '../misc/errors.js';
 import {
 	AstNode,
 	FunctionDecl,
@@ -8,104 +9,30 @@ import {
 	SourceFile,
 	StatementNode,
 	TyLabel
-} from './ast.js';
+} from '../syntax/tools.js';
 import * as builtins from './builtins.js';
-import { UguisuError } from './misc/errors.js';
-
-export class Analyzer {
-	env: AnalysisEnv;
-	symbolTable: Map<AstNode, Symbol>;
-
-	constructor() {
-		this.env = new AnalysisEnv();
-		this.symbolTable = new Map();
-	}
-
-	analyze(sourceFile: SourceFile) {
-		builtins.setDeclarations(this.env);
-		analyze({ env: this.env, symbolTable: this.symbolTable }, sourceFile);
-	}
-}
-
-export class AnalysisEnv {
-	private layers: Map<string, Symbol>[];
-
-	constructor(baseEnv?: AnalysisEnv) {
-		if (baseEnv != null) {
-			this.layers = [...baseEnv.layers];
-		} else {
-			this.layers = [new Map()];
-		}
-	}
-
-	set(name: string, symbol: Symbol) {
-		this.layers[0].set(name, symbol);
-	}
-
-	get(name: string): Symbol | undefined {
-		for (const layer of this.layers) {
-			const symbol = layer.get(name);
-			if (symbol != null) {
-				return symbol;
-			}
-		}
-		return undefined;
-	}
-
-	enter() {
-		this.layers.unshift(new Map());
-	}
-
-	leave() {
-		if (this.layers.length <= 1) {
-			throw new UguisuError('Left the root layer.');
-		}
-		this.layers.shift();
-	}
-}
-
-export type Symbol = FunctionSymbol | NativeFnSymbol | VariableSymbol | ExprSymbol;
-
-export type FunctionSymbol = {
-	kind: 'FnSymbol',
-	defined: boolean,
-	params: { name: string, ty: Type }[],
-	returnTy: Type,
-	/** for wasm */
-	vars: FnVar[],
-};
-
-export type NativeFnSymbol = {
-	kind: 'NativeFnSymbol',
-	params: { name: string, ty: Type }[],
-	returnTy: Type,
-};
-
-export function newNativeFnSymbol(params: { name: string, ty: Type }[], returnTy: Type): NativeFnSymbol {
-	return { kind: 'NativeFnSymbol', params, returnTy };
-}
-
-export type VariableSymbol = {
-	kind: 'VariableSymbol',
-	defined: boolean,
-	ty?: Type,
-};
-
-export type ExprSymbol = {
-	kind: 'ExprSymbol',
-	ty: Type,
-};
-
-export type Type = 'void' | 'number' | 'bool' | 'string' | 'function';
-
-export type FnVar = { name: string, isParam: boolean, ty?: Type };
+import {
+	AnalysisEnv,
+	assertType,
+	dispatchError,
+	FnVar,
+	FunctionSymbol,
+	Symbol,
+	Type,
+	VariableSymbol
+} from './tools.js';
 
 type Context = {
-	symbolTable: Map<AstNode, Symbol>,
 	env: AnalysisEnv,
+	symbolTable: Map<AstNode, Symbol>,
 };
 
-function analyze(ctx: Context, source: SourceFile) {
+export function analyze(source: SourceFile, env: AnalysisEnv, symbolTable: Map<AstNode, Symbol>) {
+	const ctx: Context = {
+		env,
+		symbolTable,
+	};
+	builtins.setDeclarations(ctx.env);
 	for (const n of source.funcs) {
 		setDeclaration(ctx, n);
 	}
@@ -113,15 +40,6 @@ function analyze(ctx: Context, source: SourceFile) {
 		validateFunc(ctx, n);
 	}
 	// console.log(ctx.symbolTable);
-}
-
-function assertType(actual: Type, expected: Type, errorNode: AstNode) {
-	if (actual == 'void') {
-		dispatchError(`A function call that does not return a value cannot be used as an expression.`, errorNode);
-	}
-	if (actual != expected) {
-		dispatchError(`type mismatched. expected \`${expected}\`, found \`${actual}\``, errorNode);
-	}
 }
 
 function setDeclaration(ctx: Context, node: AstNode) {
@@ -416,12 +334,4 @@ function lookupSymbolWithNode(ctx: Context, node: AstNode): Symbol {
 		dispatchError('unknown identifier.', node);
 	}
 	return symbol;
-}
-
-function dispatchError(message: string, errorNode?: AstNode): never {
-	if (errorNode != null) {
-		throw new UguisuError(`${message} (${errorNode.pos[0]}:${errorNode.pos[1]})`);
-	} else {
-		throw new UguisuError(message);
-	}
 }
