+++ test/run.ts
@@ -1,41 +1,9 @@
 import assert from 'assert';
-import { Analyzer } from '../src/lib/analyze.js';
-import { SourceFile } from '../src/lib/ast.js';
-import { Parser } from '../src/lib/parse.js';
-import { Runner } from '../src/lib/run.js';
+import { Uguisu } from '../src/lib/index.js';
 
 function runTest(sourceCode: string) {
-	const parser = new Parser();
-
-	let ast: SourceFile;
-	try {
-		ast = parser.parse(sourceCode, 'test.ug');
-	} catch (err) {
-		if (err instanceof Error) {
-			throw new Error(`Parsing Error: ${err.message}`);
-		}
-		throw err;
-	}
-
-	const analyzer = new Analyzer();
-	try {
-		analyzer.analyze(ast);
-	} catch (err) {
-		if (err instanceof Error) {
-			throw new Error(`Syntax Error: ${err.message}`);
-		}
-		throw err;
-	}
-
-	const runner = new Runner({});
-	try {
-		runner.run(ast);
-	} catch (err) {
-		if (err instanceof Error) {
-			throw new Error(`Runtime Error: ${err.message}`);
-		}
-		throw err;
-	}
+	const uguisu = new Uguisu();
+	uguisu.runCode(sourceCode, 'test.ug');
 }
 
 // variable + number literal
